add donate button back
