+++ app/leaderboard/page.tsx
@@ -6,7 +6,9 @@ const DuneDataDisplay = () => {
   return (
     <div className="flex flex-col items-center py-8 lg:p-8">
       <h2 className="text-thin flex text-center text-4xl">DONATIONS</h2>
-      <DonateButton />
+      <div className="my-4">
+        <DonateButton />
+      </div>
       <SupportedTokens />
       <Leaderboard />
     </div>
